+++ eras/babbage/impl/test/Test/Cardano/Ledger/Babbage/BinarySpec.hs
@@ -7,15 +7,14 @@ module Test.Cardano.Ledger.Babbage.BinarySpec (spec) where
 
 import Cardano.Ledger.Alonzo.TxWits (Redeemers, TxDats)
 import Cardano.Ledger.Babbage
-import Cardano.Ledger.Block (Block)
+import Cardano.Ledger.Block (Block (Block))
 import Cardano.Protocol.Crypto (StandardCrypto)
 import qualified Cardano.Protocol.Praos.BlockHeader as Praos
+import Test.Cardano.Ledger.Alonzo.Arbitrary (genSmallAlonzoBlockBody)
 import Test.Cardano.Ledger.Alonzo.Binary.RoundTrip (roundTripAlonzoCommonSpec)
 import Test.Cardano.Ledger.Babbage.Arbitrary ()
 import Test.Cardano.Ledger.Babbage.Era ()
 import Test.Cardano.Ledger.Babbage.TreeDiff ()
-import Test.Cardano.Ledger.Binary (decoderEquivalenceSpec)
-import Test.Cardano.Ledger.Binary.RoundTrip (roundTripCborSpec)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Binary as Binary (
   decoderEquivalenceCoreEraTypesSpec,
@@ -32,17 +31,15 @@ spec :: Spec
 spec = do
   describe "RoundTrip" $ do
     roundTripAlonzoCommonSpec @BabbageEra
-    roundTripCborSpec @(Praos.HeaderBody StandardCrypto)
-    roundTripCborSpec @(Praos.Header StandardCrypto)
     prop "Block (Praos.Header)" $
       withMaxSuccess 25 $
-        conjoin
-          [ roundTripEraExpectation @BabbageEra @(Block (Praos.Header StandardCrypto) BabbageEra)
-          , roundTripAnnEraExpectation @BabbageEra @(Block (Praos.Header StandardCrypto) BabbageEra)
-          ]
+        forAll (Block <$> arbitrary <*> genSmallAlonzoBlockBody) $ \block ->
+          conjoin
+            [ roundTripEraExpectation @BabbageEra @(Block (Praos.Header StandardCrypto) BabbageEra) block
+            , roundTripAnnEraExpectation @BabbageEra @(Block (Praos.Header StandardCrypto) BabbageEra) block
+            ]
   describe "DecCBOR instances equivalence" $ do
     Binary.decoderEquivalenceCoreEraTypesSpec @BabbageEra
     decoderEquivalenceEraSpec @BabbageEra @(TxDats BabbageEra)
     decoderEquivalenceEraSpec @BabbageEra @(Redeemers BabbageEra)
-    decoderEquivalenceSpec @(Praos.Header StandardCrypto) minBound maxBound
   Binary.txSizeSpec @BabbageEra