+++ eras/conway/impl/test/Main.hs
@@ -3,12 +3,13 @@
 
 module Main where
 
-import Cardano.Ledger.Block (Block)
+import Cardano.Ledger.Block (Block (Block))
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Tx (tierRefScriptFee)
 import Cardano.Protocol.Crypto (StandardCrypto)
 import qualified Cardano.Protocol.Praos.BlockHeader as Praos
+import Test.Cardano.Ledger.Alonzo.Arbitrary (genSmallAlonzoBlockBody)
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Conway.GenesisSpec as Genesis
@@ -47,10 +48,11 @@ main = ledgerEraTestMain @ConwayEra $ do
     describe "RoundTrip" $
       prop "Block (Praos.Header)" $
         withMaxSuccess 25 $
-          conjoin
-            [ roundTripEraExpectation @ConwayEra @(Block (Praos.Header StandardCrypto) ConwayEra)
-            , roundTripAnnEraExpectation @ConwayEra @(Block (Praos.Header StandardCrypto) ConwayEra)
-            ]
+          forAll (Block <$> arbitrary <*> genSmallAlonzoBlockBody) $ \block ->
+            conjoin
+              [ roundTripEraExpectation @ConwayEra @(Block (Praos.Header StandardCrypto) ConwayEra) block
+              , roundTripAnnEraExpectation @ConwayEra @(Block (Praos.Header StandardCrypto) ConwayEra) block
+              ]
   describe "Various tests for functions defined in Conway" $ do
     prop "tierRefScriptFee is a linear function when growth is 1" $ \(Positive sizeIncrement) baseFee (NonNegative size) ->
       tierRefScriptFee 1 sizeIncrement baseFee size