+++ eras/conway/impl/test/Test/Cardano/Ledger/Conway/Binary/CddlSpec.hs
@@ -10,6 +10,7 @@ module Test.Cardano.Ledger.Conway.Binary.CddlSpec (spec) where
 import Cardano.Ledger.Allegra.Scripts
 import Cardano.Ledger.Alonzo.Scripts (CostModels)
 import Cardano.Ledger.Alonzo.TxWits (Redeemers)
+import Cardano.Ledger.Block (Block (Block))
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Governance (GovAction, ProposalProcedure, VotingProcedure)
 import Cardano.Ledger.Conway.HuddleSpec (conwayCDDL)
@@ -17,7 +18,7 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Plutus.Data (Data, Datum)
 import Cardano.Protocol.Crypto (StandardCrypto)
 import qualified Cardano.Protocol.Praos.BlockHeader as Praos
-import Test.Cardano.Ledger.Alonzo.Arbitrary (genNonEmptyRedeemers)
+import Test.Cardano.Ledger.Alonzo.Arbitrary (genNonEmptyRedeemers, genSmallAlonzoBlockBody)
 import Test.Cardano.Ledger.Binary.Cuddle (
   huddleAntiCborSpec,
   huddleDecoderEquivalenceSpec,
@@ -45,6 +46,9 @@ genPraosHeaderBody = do
   hb <- arbitrary
   pv <- genEraProtVer @ConwayEra
   pure hb {Praos.hbProtVer = pv}
+
+genPraosBlock :: Gen (Block (Praos.Header StandardCrypto) ConwayEra)
+genPraosBlock = Block <$> genPraosHeader <*> genSmallAlonzoBlockBody
 
 spec :: Spec
 spec = do
@@ -138,6 +142,7 @@ spec = do
       huddleRoundTripGenValidate @(Praos.Header StandardCrypto) genPraosHeader v "header"
       huddleRoundTripCborSpec @(Praos.HeaderBody StandardCrypto) v "header_body"
       huddleRoundTripGenValidate @(Praos.HeaderBody StandardCrypto) genPraosHeaderBody v "header_body"
+      huddleRoundTripGenValidate @(Block (Praos.Header StandardCrypto) ConwayEra) genPraosBlock v "block"
       describe "DecCBOR instances equivalence via CDDL" $ do
         huddleDecoderEquivalenceSpec @(Praos.Header StandardCrypto) v "header"
         huddleDecoderEquivalenceSpec @(TxBody TopTx ConwayEra) v "transaction_body"